Smooth Running of UI
Add Feature for app to load quickly and work smoothly, so I can enjoy reading the news without interruptions. Also easy-to-use signup and login process, so I can save my preferences and access customized content. And app to feel confident that my personal data and login credentials are safe and protected. As a developer, I want the app to be well-tested and free from bugs to ensure a smooth user experience.